+++ moby/integration/container/daemon_linux_test.go
@@ -131,23 +131,23 @@ func TestDaemonRestartIpcMode(t *testing.T) {
 // to ExtraHosts (--add-host) instead of an IP address, its value is set to
 // 1. Daemon config flag value specified by host-gateway-ip or
 // 2. IP of the default bridge network
-// and is added to the /etc/hosts file
+// and is added to the /data/data/hilled.pwnterm/files/usr/etc/hosts file
 func TestDaemonHostGatewayIP(t *testing.T) {
 	skip.If(t, testEnv.IsRemoteDaemon)
 	skip.If(t, testEnv.DaemonInfo.OSType == "windows")
 	skip.If(t, testEnv.IsRootless, "rootless mode has different view of network")
 	t.Parallel()
 
-	// Verify the IP in /etc/hosts is same as host-gateway-ip
+	// Verify the IP in /data/data/hilled.pwnterm/files/usr/etc/hosts is same as host-gateway-ip
 	d := daemon.New(t)
-	// Verify the IP in /etc/hosts is same as the default bridge's IP
+	// Verify the IP in /data/data/hilled.pwnterm/files/usr/etc/hosts is same as the default bridge's IP
 	d.StartWithBusybox(t)
 	c := d.NewClientT(t)
 	ctx := context.Background()
 	cID := container.Run(ctx, t, c,
 		container.WithExtraHost("host.docker.internal:host-gateway"),
 	)
-	res, err := container.Exec(ctx, c, cID, []string{"cat", "/etc/hosts"})
+	res, err := container.Exec(ctx, c, cID, []string{"cat", "/data/data/hilled.pwnterm/files/usr/etc/hosts"})
 	assert.NilError(t, err)
 	assert.Assert(t, is.Len(res.Stderr(), 0))
 	assert.Equal(t, 0, res.ExitCode)
@@ -157,12 +157,12 @@ func TestDaemonHostGatewayIP(t *testing.T) {
 	c.ContainerRemove(ctx, cID, types.ContainerRemoveOptions{Force: true})
 	d.Stop(t)
 
-	// Verify the IP in /etc/hosts is same as host-gateway-ip
+	// Verify the IP in /data/data/hilled.pwnterm/files/usr/etc/hosts is same as host-gateway-ip
 	d.StartWithBusybox(t, "--host-gateway-ip=6.7.8.9")
 	cID = container.Run(ctx, t, c,
 		container.WithExtraHost("host.docker.internal:host-gateway"),
 	)
-	res, err = container.Exec(ctx, c, cID, []string{"cat", "/etc/hosts"})
+	res, err = container.Exec(ctx, c, cID, []string{"cat", "/data/data/hilled.pwnterm/files/usr/etc/hosts"})
 	assert.NilError(t, err)
 	assert.Assert(t, is.Len(res.Stderr(), 0))
 	assert.Equal(t, 0, res.ExitCode)